@@ -5,7 +5,7 @@
   "keywords": ["modx", "modx evo", "modx evolution", "ddgetdocuments"],
   "license": "MIT",
   "require": {
-    "php": ">=5.3.0",
+    "php": ">=5.4.0",
     "dd/modxevo-library-ddtools": ">=0.14",
     "composer/installers": "~1.0"
   },
