@@ -3,10 +3,12 @@
  * ddGetDocuments
  * @version 0.6 (2016-10-10)
  * 
- * A snippet for fetching and parsing resources from the document tree by a custom rule.
+ * @desc A snippet for fetching and parsing resources from the document tree by a custom rule.
+ * 
+ * @uses PHP >= 5.4.
  * 
  * @param $provider {'parent'|'select'} — Name of the provider that will be used to fetch documents. Default: 'parent'.
- * @param $providerParams {string_query} — Parameters to be passed to the provider. The parameter must be set as a query string,
+ * @param $providerParams {string_queryFormated} — Parameters to be passed to the provider. The parameter must be set as a query string,
  * When $provider == 'parent' =>
  * @param $providerParams['parentIds'] {array|string_commaSepareted} — Parent IDs. Default: 0.
  * @param $providerParams['depth'] {integer} — Depth of children documents search. Default: 1.
@@ -20,7 +22,7 @@
  * @param $orderBy {string} — A string representing the sorting rule. Default: '`id` ASC'.
  * 
  * @param $outputFormat {'string'|'json'|'raw'} — Format of the output. Default: 'string'.
- * @param $outputFormatParams {string_query} — Parameters to be passed to the specified formatter. The parameter must be set as a query string,
+ * @param $outputFormatParams {string_queryFormated} — Parameters to be passed to the specified formatter. The parameter must be set as a query string,
  * When $outputFormat == 'string' =>
  * @param $outputFormatParams['itemTpl'] {string_chunkName} — Item template. Available placeholders: [+any field or tv name+], [+any of extender placeholders+]. @required
  * @param $outputFormatParams['itemTplFirst'] {string_chunkName} — First item template. Available placeholders: [+any field or tv name+], [+any of extender placeholders+].
@@ -34,7 +36,7 @@
  * 
  * @param $extenders {string_commaSeparated} — Comma-separated string determining which extenders should be applied to the snippet.
  * Be aware that the order of extender names can affect the output.
- * @param $extendersParams {string_query} — Parameters to be passed to their corresponding extensions. The parameter must be set as a query string,
+ * @param $extendersParams {string_queryFormated} — Parameters to be passed to their corresponding extensions. The parameter must be set as a query string,
  * When $extenders == 'pagination' =>
  * @param $extendersParams['wrapperTpl'] {string_chunkName} — Chunk to be used to output the pagination. @required
  * @param $extendersParams['pageTpl'] {string_chunkName} — Chunk to be used to output pages within the pagination. @required
@@ -44,6 +46,8 @@
  * @param $extendersParams['previousTpl'] {string_chunkName} — Chunk to be used to output the navigation block to the previous page. @required
  * @param $extendersParams['previousOffTpl'] {string_chunkName} — Chunk to be used to output the navigation block to the previous page if there are no more pages before. @required
  * e.g. $extendersParams = 'wrapperTpl=pagination&nextTpl=pagination_next&previousTpl=pagination_previous&nextOffTpl=pagination_nextOff&previousOffTpl=pagination_previousOff&pageTpl=pagination_page&currentPageTpl=pagination_page_current'.
+ * 
+ * @copyright 2015–2016 DivanDesign {@link http://www.DivanDesign.biz }
  **/
 
 global $modx;
@@ -86,7 +90,7 @@ if(class_exists($providerClass)){
 	//Prepare provider params
 	parse_str($providerParams, $providerParamsArray);
 	
-	$extendersNamesArray = array();
+	$extendersNamesArray = [];
 	
 	if($extenders != ''){
 		$extendersNamesArray = explode(',', $extenders);
@@ -99,15 +103,15 @@ if(class_exists($providerClass)){
 		//like [extenderName => [extenderParameter_1, extenderParameter_2, ...]]
 		if(count($extendersNamesArray) === 1){
 			if(!isset($extendersParamsArray[$extendersNamesArray[0]])){
-				$extendersParamsArray = array(
+				$extendersParamsArray = [
 					$extendersNamesArray[0] => $extendersParamsArray
-				);
+				];
 			}
 		}else{
 			//Make sure that for each extender there is an item in $extendersParamsArray 
 			foreach($extendersNamesArray as $extenderName){
 				if(!isset($extendersParamsArray[$extenderName])){
-					$extendersParamsArray[$extenderName] = array();
+					$extendersParamsArray[$extenderName] = [];
 				}
 			}
 		}
@@ -120,20 +124,20 @@ if(class_exists($providerClass)){
 	
 	$input = new \ddGetDocuments\Input(
 		//Snippet params
-		array(
+		[
 			'offset' => $offset,
 			'total' => $total,
 			'orderBy' => $orderBy,
 			'filter' => $filter,
 			'fieldDelimiter' => $fieldDelimiter
-		),
+		],
 		$providerParamsArray,
 		$extendersParamsArray,
 		$outputFormatParamsArray
 	);
 	
 	//Extenders storage
-	$extendersStorage = array();
+	$extendersStorage = [];
 	
 	//Iterate through all extenders to create their instances
 	foreach($extendersNamesArray as $extenderName){
